Test method testExceptionCoordinateOrQuerySortRequired() replaced to MarkDuplicatesExceptionTest class, and MarkDuplicatesExceptionTest class removed.
